Added scripts/lldb-commands.sh
This script contains several commands that are intended to aid lldb
development workflow.  At the moment it depends on some newer shell
constructs and bash-isms (lightly).  pushd/popd, local, eval, etc.
If we use more than bash, I can classicalize it to support a wider
range of shells.

End-user commands
=================

reload_lldb_commands: reloads the lldb-commands.sh script regardless of
its location on your filesystem. Useful when implementing and testing
changes to lldb-commands.sh itself.

pull_llvm: finds the llvm directory somewhere in the parent directory
chain, and executes a 'git pull origin' on it. Right now it assumes that
a 'git pull origin' is sufficient to do the right thing. I'm operating
out of the assumption that our workflow for the short term will be heavy
modifications to the lldb repository but will use llvm and clang
repositories directly from master.

pull_clang: similar to pull_llvm, but finds the clang directory (assumed
to be under the llvm directory as tools/clang). It does not require the
user to be within the clang tree; however, it does require the user to
be somewhere within the llvm tree

pull_lldb_rebase: this one does a bunch of operations.  The intended
usage is to support a developer who is working on a non-master branch
of lldb, but that shares a common heritage with the lldb master branch.
The user has made changes, some of which may be checked into the
local branch (or one of its ancestors), but are not upstreamed to lldb
yet.  The user wants to "sync up with upstream," preserving any of
the as-yet-not-upstreamed changes checked in (or not checked in yet
but within the working tree).  Here are the operations that are
performed:
* add any untracked files to the current branch
* stash the current branch
* change to the master branch
* do a 'git pull origin' on the master branch
* change back to the initial branch
* do a 'git rebase master' to push local changes to the end of the
  commit history,
* replay any stashed changes on top of the initial branch

This command does error checking along the way.  It will stop
doing potentially destructive operations the first point that
an error is hit (e.g. git pull origin, git rebase master, or
git stash pop).

cfglldb: runs '../llvm/configure --enable-cxx11 \
--prefix=`pwd`/../$INSTALL_DIR

The INSTALL_DIR var is optionally passed in via the first
argument to the method, defaulting to 'install' if not specified.

mklog: runs "make $@ 2>&1 | tee make.log"

mkilog: runs "make $@ install 2>&1 | tee make_install.log"

Internal commands
=================

find_dir_parent_chain: given (1) a variable name to use as a return
variable (may be global or local), and (2) a directory path fragment to
find at the current directory level or one of its parents, returns the
path to the directory within the parent chain that contains the given
directory path fragment as a suffix. Returns 0 on success, 1 on failure.
On success, the variable named by the first argument will contain the
resulting absolute directory path.

git_pull: given (1) a directory in which to run the git pull,
(2) an optional remote repository to pull from (default: origin),
(3) an optional branch/branch-mapping to specify (default: ''), execute
a 'git pull {remote} {branch/branch-mapping}'. Returns 0 on success, 1
on failure. Working directory is unmodified on exit.