Renamed cfglldb to configure_lldb, enhanced.
The configure_lldb command now follows this usage:

usage: configure_lldb [BUILD_DIR] [INSTALL_DIR]
where BUILD_DIR defaults to build and INSTALL_DIR defaults to install.

The command first searches up the tree for the parent of the
llvm source.  It assumes the BUILD_DIR AND INSTALL_DIR are
relative to this parent directory (i.e. siblings of the llvm
source tree).

It checks if a 'clang' exe is in the path. If so, it fails.

It checks if the build dir already exists.  If so, it fails.

It checks if the install dir already exists.  If so, it fails.

Otherwise, it creates the build dir, runs configure within it,
setting it to install in the INSTALL_DIR.