Added local python and system libedit support to configure.
lldb_configure.py changes:

* Added --with-python-dir ROOT_DIR flag, which defaults to the
  environment variable contents of LLDB_PYTHON_BASE_DIR if specifed;
  otherwise, uses the python available in the path (typically the
  system python).

* Added --stock-libedit flag, which defaults to being false.  When
  specified, it turns off our tools/libedit version of libedit
  from being used.

* Alphabetized the ordering of lldb config flag specification by long
  name. This was always the intent, but somehow I let them get out of
  order.

Added lldb_setenv_local_python.sh. This file is meant to be
sourced (e.g. '. ~/lldb/tools/scripts/lldb_setenv_local_python.sh') to
set several environment variables that must be present for a local
python to be used in the configuration, build and execution of lldb with
the custom python.  When sourced, this will set the environment variable
that will trigger lldb_configure.py to configure lldb with the
custom local python.  Right now it assumes the python is available
at the root /usr/local/python/python-release.

Change-Id: I2c9cce0fa31592108e1660c86862cdbcee7c0db5