Added gpy command to run gpylint --rcfile={SCRIPT_DIR}/support/pylintrc $@
This is a convenience alias-like call to simplify calling gpylint with
our modified pylintrc that suppresses some Google3 style recommendations
that don't make sense for running outside of prod.

gpy does require that $HOME/lldb/tools/scripts/support/pylintrc exists.
If it doesn't, it prints an error message and skips the command.

Change-Id: I0cf3af15c22d4f813745aef59abaa2c37ca3730c