Remove cmake LLVM_ENABLE_CXX11 flag.
LLVM for a while now has been broadcasting that the default would change
to always assume C++11 support. This has happened since I now get a
warning about an unused flag passed to cmake. This change removes the
use of the flag (and thus eliminates the warning) when doing an
'lldb_configure.py -c/--cmake' operation.

Change-Id: Id5e9abde6433585715d8cbf4e549796d285fd623