Add a script that can set/unset llgs commandline args when executed by lldb.
When using remote debugging for local processes, lldb starts up llgs manually.
lldb looks for certain environment variables to see if anything extra should
be passed to llgs when it is started.  This script sets some variables to turn
on logging to /tmp/llgs_packets.log and /tmp/llgs_process.log.

I have found that turning on this logging enables one of the races to disappear
on my end (https://github.com/tfiala/lldb/issues/50) pretty much all the time.

Checking this in so that Alex and/or Shawn may be able to use it to play around
with timing.  I am going to ignore the race for now since this script gets me
past that issue.  Forging ahead...